extracting float deconstruct code into own component
